backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.9.2
email-validator==2.2.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
alembic==1.14.0
websockets==13.1
celery==5.4.0
httpx==0.27.2
google-generativeai==0.8.4
python-dotenv==1.0.1
